スピーチモデル関連のクラスを整理・追加
`SpeechCustomizationApiAccess.cs` にスピーチモデルの API 説明を追加し、メソッドコメントを更新しました。スピーチモデルのリクエストおよびレスポンスモデルのクラス定義は削除され、新しいファイルに移動されました。

新たに `ApiSpeechModelPropertiesModel.cs`、`ApiSpeechModelRequestModel.cs`、`ApiSpeechModelResponseModel.cs` を追加し、それぞれのクラスに必要なプロパティを定義しました。
