+++ VideoIndexerAccessCore/VideoIndexerClient/ApiAccess/SpeechCustomizationApiAccess.cs
@@ -154,6 +154,8 @@ namespace VideoIndexerAccessCore.VideoIndexerClient.ApiAccess
 
         /// <summary>
         /// API からスピーチモデル作成の JSON データを取得します。
+        /// Create Speech Model
+        /// https://api-portal.videoindexer.ai/api-details#api=Operations&operation=Create-Speech-Model
         /// </summary>
         /// <param name="location">Azure のリージョン</param>
         /// <param name="accountId">アカウント ID</param>
@@ -199,6 +201,8 @@ namespace VideoIndexerAccessCore.VideoIndexerClient.ApiAccess
 
         /// <summary>
         /// API を呼び出してスピーチモデルを作成します。
+        /// Create Speech Model
+        /// https://api-portal.videoindexer.ai/api-details#api=Operations&operation=Create-Speech-Model
         /// </summary>
         /// <param name="location">Azure のリージョン</param>
         /// <param name="accountId">アカウント ID</param>
@@ -213,6 +217,8 @@ namespace VideoIndexerAccessCore.VideoIndexerClient.ApiAccess
 
         /// <summary>
         /// JSON をパースして ApiSpeechModelResponseModel オブジェクトに変換します。
+        /// Create Speech Model
+        /// https://api-portal.videoindexer.ai/api-details#api=Operations&operation=Create-Speech-Model
         /// </summary>
         /// <param name="json">API から取得した JSON 文字列</param>
         /// <returns>パースした ApiSpeechModelResponseModel オブジェクト、エラー時は null</returns>
@@ -227,43 +233,6 @@ namespace VideoIndexerAccessCore.VideoIndexerClient.ApiAccess
                 _logger.LogError($"JSON parsing error: {ex.Message}");
                 return null;
             }
-        }
-
-        /// <summary>
-        /// スピーチモデルの作成リクエストを表します。
-        /// </summary>
-        public class ApiSpeechModelRequestModel
-        {
-            public string displayName { get; set; } = string.Empty; // モデルの表示名
-            public string locale { get; set; } = string.Empty; // 言語ロケール
-            public List<string> datasets { get; set; } = new List<string>(); // 関連データセットの ID リスト
-            public string description { get; set; } = string.Empty; // モデルの説明
-            public Dictionary<string, string> customProperties { get; set; } = new Dictionary<string, string>(); // カスタムプロパティ
-        }
-
-        /// <summary>
-        /// スピーチモデルの API レスポンスモデル
-        /// </summary>
-        public class ApiSpeechModelResponseModel
-        {
-            public string? id { get; set; } // モデルの ID
-            public SpeechModelProperties? properties { get; set; } // モデルのプロパティ
-            public string? displayName { get; set; } // モデルの表示名
-            public string? description { get; set; } // モデルの説明
-            public string? locale { get; set; } // 言語ロケール
-            public List<string>? datasets { get; set; } // データセットの ID リスト
-            public int status { get; set; } // モデルのステータス (0:None, 1:Waiting, 2:Processing, 3:Complete, 4:Failed)
-            public string? lastActionDateTime { get; set; } // 最終更新日時
-            public string? createdDateTime { get; set; } // 作成日時
-            public Dictionary<string, string>? customProperties { get; set; } // カスタムプロパティ
-        }
-
-        /// <summary>
-        /// スピーチモデルのプロパティ情報
-        /// </summary>
-        public class SpeechModelProperties
-        {
-            public string? error { get; set; } // エラーメッセージ (存在する場合)
         }
     }
 }
