+++ VideoIndexerAccessCore/VideoIndexerClient/ApiModel/ApiSpeechModelPropertiesModel.cs
@@ -1,0 +1,10 @@
+﻿namespace VideoIndexerAccessCore.VideoIndexerClient.ApiModel
+{
+    /// <summary>
+    /// スピーチモデルのプロパティ情報
+    /// </summary>
+    public class ApiSpeechModelPropertiesModel
+    {
+        public string? error { get; set; } // エラーメッセージ (存在する場合)
+    }
+}
